Added different modulation schemes on the dashboard

+++ BPSK.py
@@ -17,17 +17,18 @@ from scipy.stats import norm
 # Tb = 0.01
 # Eb = 0.001
 
+
 def modulate(msg, Eb, Tb, f_c, f_s):
     modulated_signal = []
-    t = np.linspace(0, Tb, Tb*f_s)
+    t = np.linspace(0, Tb, int(Tb * f_s))
     for i in msg:
-        s = np.sqrt(2*Eb/Tb)*np.sin(2*np.pi*f_c*t)
+        s = np.sqrt(2 * Eb / Tb) * np.sin(2 * np.pi * f_c * t)
         if i == 0:
             s = -s
         modulated_signal.extend(s)
-    t = np.linspace(0, len(msg)*Tb, len(msg)*Tb*f_s)
+    t = np.linspace(0, len(msg) * Tb, int(len(msg) * Tb * f_s))
     modulated_signal = np.array(modulated_signal)
-    return modulated_signal    
+    return modulated_signal
 
 # def add_noise(signal, N0):
 #     N0_unit_power = 0.0004
@@ -40,23 +41,25 @@ def modulate(msg, Eb, Tb, f_c, f_s):
 #     signal_with_noise = signal + noise
 #     return signal_with_noise
 
+
 def demodulate(signal, Tb, f_c, f_s):
-    t = np.linspace(0, Tb, Tb*f_s)
-    phi = np.sqrt(2/Tb)*np.sin(2*np.pi*f_c*t)
+    t = np.linspace(0, Tb, int(Tb * f_s))
+    phi = np.sqrt(2 / Tb) * np.sin(2 * np.pi * f_c * t)
     N = len(signal) // len(t)
     signal = np.array_split(signal, N)
     received_msg = []
     for i in signal:
-        x = i*phi
-        sm = x.sum()/f_s
+        x = i * phi
+        sm = x.sum() / f_s
         if sm > 0:
             received_msg.append(1)
         else:
             received_msg.append(0)
     return received_msg
 
+
 def error_probabilities(msg, decoded_msg, Eb, N0):
-    Pb = norm.sf(np.sqrt(2*Eb/N0))
+    Pb = norm.sf(np.sqrt(2 * Eb / N0))
     # print('Theoretical Bit Error Probability:', Pb)
     Pb_pr = np.count_nonzero(msg != decoded_msg) / len(msg)
     # print('Practical Bit Error Probability:', Pb_pr)
